Less noise for incomplete antenna/mode sets
git-svn-id: https://svn.atnf.csiro.au/difx@10592 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/vex2difx/trunk/src/autobands.cpp
@@ -652,7 +652,7 @@ int AutoBands::generateOutputbandsExplicit(int Nant_min)
 		while(!newband.isComplete())
 		{
 			int startspan = -1;
-			for(int span=0; span<spans.size(); span++)
+			for(size_t span=0; span<spans.size(); span++)
 			{
 				if(spans[span].flow <= foutstart && foutstart < spans[span].fhigh)
 				{
